+++ src/ipl3checksum/frontends/sum.py
@@ -10,7 +10,7 @@ from pathlib import Path
 
 import ipl3checksum
 
-def doDetectCic(romBytes: bytes) -> int:
+def doChecksumize(romBytes: bytes) -> int:
     kind = ipl3checksum.detectCIC(romBytes)
 
     if kind is None:
@@ -18,6 +18,14 @@ def doDetectCic(romBytes: bytes) -> int:
         return 1
 
     print(f"Detected kind is '{kind.name}'")
+
+    checksum = ipl3checksum.calculateChecksum(romBytes, kind)
+    if checksum is None:
+        print(f"Unable to calculate checksum")
+        return 1
+
+    chk0, chk1 = checksum
+    print(f"Calculated checksum: {chk0:08X} {chk1:08X}")
 
     return 0
 
@@ -27,10 +35,10 @@ def processArguments(args: argparse.Namespace):
 
     romBytes = romPath.read_bytes()
 
-    exit(doDetectCic(romBytes))
+    exit(doChecksumize(romBytes))
 
 def addSubparser(subparser: argparse._SubParsersAction[argparse.ArgumentParser]):
-    parser = subparser.add_parser("detect_cic", help="Detects the CIC variant of a given rom")
+    parser = subparser.add_parser("sum", help="Calculates the ipl3 checksum of a big endian ROM.")
 
     parser.add_argument("rom_path", help="Path to a big endian ROM file", type=Path)